move vtab stuff into its on h/cpp
